Tests/LibWeb/Text/expected/async-scrolling/overflow-hidden-axis-wheel-target.txt

horizontal wheel target: non-viewport
vertical wheel target: none
